some changes to improve the handling of multiple data collector.
there was a problem that the producer when they got the address for the first dc they where not able to receive  another one.
now the producer only receive the dc address when they are about to configure. With this change the DC receives only one dc address and the bug does not occur.
it is more a work around than a fix